soma-241/feat - add Login page
