+++ pages/oauth/style.module.scss
@@ -1,7 +1,7 @@
 @import '@styles/main.scss';
 
 .oauth-loading {
-  height: 100vh;
+  height: calc(100vh - 60px);
   display: flex;
   flex-direction: column;
   align-items: center;
